requirements.txt
transformers
langchain
openai
torch
spacy
nltk
spotipy
streamlit
fastapi
uvicorn
python-dotenv
